+++ src/main/java/org/folio/ld/dictionary/label/generators/FormLabelGenerator.java
@@ -1,0 +1,21 @@
+package org.folio.ld.dictionary.label.generators;
+
+import static org.folio.ld.dictionary.PropertyDictionary.NAME;
+import static org.folio.ld.dictionary.ResourceTypeDictionary.FORM;
+import static org.folio.ld.dictionary.label.LabelHelper.getPropertyValue;
+
+import org.folio.ld.dictionary.label.LabelGenerator;
+import org.folio.ld.dictionary.model.Resource;
+
+public class FormLabelGenerator implements LabelGenerator {
+
+  @Override
+  public boolean supports(Resource resource) {
+    return resource.isOfType(FORM);
+  }
+
+  @Override
+  public String getLabel(Resource resource) {
+    return getPropertyValue(resource, NAME.getValue()).orElse("");
+  }
+}